port(qcommon,server): VM call seam — VM_Call, slot-armed game progs, botlib API handshake
VM_Call lands in vm_fns (native arm: 16-word arg pack into the
C-variadic vmMain entryPoint, currentVM/lastVM save-restore,
vm_debugLevel trace; the QVM VM_CallCompiled/Interpreted arms are dead
surface in the native-dll build -> ERR_FATAL, §20). SV_InitGameProgs
arms the SEAM-D11 GAME_SLOT (sv ctx + game_system_calls_shim) before
VM_Create, whose systemCalls now takes a C-ABI adapter widening the
legacy int frame to trampoline words — the last E0308 is gone and both
server_host todo!() arms are replaced by Raven's own Bad-game-system-
trap ERR_DROP default. SV_ShutdownGameProgs ported (GAME_SHUTDOWN,
VM_Free, gvm null). SV_BotInitBotLib drops its stale BotLib placeholder
for the real mp_engine_botlib::BotLib: debugpoly pool realloc,
GetBotLibAPI handshake, export stored on sv.botlib_export.

needsSession flagged: populating botlib_import_t's 22 C-ABI fn-ptr
fields needs a BOTLIB import slot (state-capture dual of GAME_SLOT);
table ships zero-init until that seam is designed.

mp_engine_server 195 -> 178 (65 E0425 + 98 E0599 + 9 E0433 + 4 E0609 +
2 E0432, zero E0308/E0061); qcommon/botlib 0; jampgame green. (Local
commit, push batched.)

+++ crates/mp/engine/server/src/sv_ccmds.rs
@@ -21,6 +21,7 @@ use mp_engine_qcommon::cmd_pc::Server as CmdServerSlot;
 use mp_engine_qcommon::common::opaque_slots::Ghoul2System as CmdGhoul2Slot;
 use mp_engine_qcommon::collision_world::CollisionWorld;
 use mp_engine_qcommon::common::common::Common;
+use mp_engine_qcommon::vm::VM_Call;
 use mp_engine_qcommon::cm_load::RenderModels;
 use mp_engine_qcommon::cm_load::RmManager;
 use mp_host_interface::engine_host::EngineHost;
@@ -1161,7 +1162,7 @@ pub fn SV_MapRestart_f(
                 common,
                 sv.gvm,
                 mp_abi::game::exports::MpGameExport::GAME_CLIENT_CONNECT as c_int,
-                &[i, qboolean::qfalse as isize, isBot as isize],
+                &[i, qboolean::qfalse as c_int, isBot as c_int],
             ),
         ) as *mut c_char;
         if !denied.is_null() {
